requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
numpy==1.19.1
pytz==2020.1
sqlparse==0.3.1
